Cast physicalDpiX to double

+++ unit-tests/UI_test.cpp
@@ -140,11 +140,11 @@ void TestUI::ScreenCalibrationWidget_dpi()
 {
 	Tw::UI::ScreenCalibrationWidget w;
 	QSignalSpy spy(&w, SIGNAL(dpiChanged(double)));
-	constexpr double MagicDPI = 42.;
+	constexpr double MagicDPI = 42;
 
 	QVERIFY(spy.isValid());
 
-	QCOMPARE(w.dpi(), w.physicalDpiX());
+	QCOMPARE(w.dpi(), static_cast<double>(w.physicalDpiX()));
 	QCOMPARE(spy.count(), 0);
 	w.setDpi(MagicDPI);
 	QCOMPARE(spy.count(), 1);
